fix: Cleanup report code
- Show toolbar during load
- Create ReportContentComponent -> allows memoizing PDFReport
- More robust loading code (patch for @react-pdf see
    https://github.com/diegomura/react-pdf/pull/995)
