+++ src/renderer/components/MapFilter/ViewWrapper.js
@@ -35,7 +35,6 @@ export type CommonViewProps = {
 type Props = {
   ...$Exact<CommonViewProps>,
   children: ({
-    filter: Filter,
     filteredObservations: Array<Observation>,
     onClickObservation: (observationId: string, imageIndex?: number) => void,
     getPreset: Observation => PresetWithAdditionalFields,
